Move hard delete methods to methods/file_revision.rs
